+++ src/main/java/cech12/usefulhats/UsefulHatsMod.java
@@ -5,10 +5,13 @@ import cech12.usefulhats.compat.CuriosMod;
 import cech12.usefulhats.config.CommonConfig;
 import cech12.usefulhats.config.ServerConfig;
 import cech12.usefulhats.init.ModItems;
+import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.ItemStack;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
+import net.minecraftforge.event.CreativeModeTabEvent;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.DistExecutor;
@@ -30,13 +33,7 @@ public class UsefulHatsMod {
 
     public static final String MOD_ID = "usefulhats";
 
-    public static final CreativeModeTab ITEM_GROUP = new CreativeModeTab("usefulhats") {
-        @Override
-        @NotNull
-        public ItemStack makeIcon() {
-            return new ItemStack(ModItems.STOCKING_CAP.get());
-        }
-    };
+    public static CreativeModeTab ITEM_GROUP;
 
     public UsefulHatsMod() {
         IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
@@ -64,4 +61,14 @@ public class UsefulHatsMod {
         ModItems.setupClient();
     }
 
+    @SubscribeEvent
+    public static void registerTabs(CreativeModeTabEvent.Register event) {
+        ITEM_GROUP = event.registerCreativeModeTab(new ResourceLocation(MOD_ID, "main_tab"), builder -> builder
+                .icon(() -> new ItemStack(ModItems.STOCKING_CAP.get()))
+                .title(Component.translatable("tabs.usefulhats.main_tab"))
+                .displayItems((featureFlags, output, hasOp) -> {
+                    ModItems.ITEMS.getEntries().forEach(itemRegistryObject -> output.accept(itemRegistryObject.get()));
+                })
+        );
+    }
 }